Refactor construction of bib dicts into method
I am converging on using a dict as a container for all the data I need
for the "test" functionality. The idea is to iterate over all of the
fully-qualified pathnames that come in from the command line. For each,
I gather all the information that could be needed at the same time and
collect that information in a dictionary. Then, I create a list out of
all those dictionaries.

In this architecture, the construction of these dictionaries should be
done by a single method, which was created as of this commit.
Furthermore, there were some superfluous methods that were deleted.